Write & Read Duplicate Inine Schemas Correctly
If a user defined record is repeated within a schema it is only inlined
once. The second reference to that type is simply specified by name. Before
this change, the name would not be written out or parsed correctly
because the code did not properly handle (de)-serialization of
RecordReferenceSchemas.

This change includes tests confirming the desired behavior.